Relocated the components to better clarify admin and guest
