+++ app/Http/Controllers/Admin/AnimalController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
 use App\Http\Requests\StoreAnimalRequest;
 use App\Models\Animal;
@@ -12,6 +12,7 @@ use Inertia\Inertia;
 use Inertia\Response;
 use Spatie\QueryBuilder\AllowedFilter;
 use Spatie\QueryBuilder\QueryBuilder;
+use App\Http\Controllers\Controller;
 
 class AnimalController extends Controller
 {
