+++ app/Http/Controllers/PetsController.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace App\Http\Controllers;
+use Illuminate\Http\Request;
+
+class PetsController extends Controller
+{
+    // Display a listing of the resource.
+    public function index()
+    {
+        //
+    }
+
+    // Show the form for creating a new resource.
+    public function create()
+    {
+        //
+    }
+
+    // Store a newly created resource in storage.
+    public function store(Request $request)
+    {
+        //
+    }
+
+    // Display the specified resource.
+    public function show($id)
+    {
+        //
+    }
+
+    // Show the form for editing the specified resource.
+    public function edit($id)
+    {
+        //
+    }
+
+    // Update the specified resource in storage.
+    public function update(Request $request, $id)
+    {
+        //
+    }
+
+    // Remove the specified resource from storage.
+    public function destroy($id)
+    {
+        //
+    }
+}
