+++ routes/web.php
@@ -3,6 +3,7 @@
 declare(strict_types=1);
 
 use App\Http\Controllers\DashboardController;
+use App\Http\Controllers\PetsController;
 use App\Http\Controllers\ProfileController;
 use Illuminate\Foundation\Application;
 use Illuminate\Support\Facades\Route;
@@ -14,8 +15,11 @@ Route::get('/', function () {
         'canRegister' => Route::has('register'),
         'laravelVersion' => Application::VERSION,
         'phpVersion' => PHP_VERSION,
+        'appName' => config('app.name')
     ]);
 });
+
+Route::get('/pets', [PetsController::class, 'index'])->name('pets.index');
 
 Route::get('/dashboard', DashboardController::class)->middleware(['auth', 'verified'])->name('dashboard');
 
@@ -25,7 +29,7 @@ Route::middleware('auth')->group(function () {
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 });
 
-Route::resource('animals', \App\Http\Controllers\AnimalController::class)
+Route::resource('animals', \App\Http\Controllers\Admin\AnimalController::class)
     ->middleware(['auth', 'verified']);
 
 require __DIR__.'/auth.php';
